# docs/swagger.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Book Cloud API",
        "description": "Cloud API for books recommendation and statistics",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/ratings": {
            "get": {
                "tags": [
                    "Ratings"
                ],
                "summary": "List ratings",
                "operationId": "get_ratings",
                "parameters": [
                    {
                        "name": "page_number",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "title": "Page Number"
                        },
                        "description": "Optional page number for paginated results."
                    },
                    {
                        "name": "page_size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "title": "Page Size"
                        },
                        "description": "Optional page size for paginated results."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Rating"
                                    },
                                    "title": "Ratings Response"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/rating/{book_isbn}": {
            "post": {
                "tags": [
                    "Ratings"
                ],
                "summary": "Add rating to book",
                "operationId": "rating_add_post",
                "parameters": [
                    {
                        "name": "book_isbn",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Book ISBN",
                            "examples": [
                                "9780439554930"
                            ]
                        },
                        "description": "ISBN of the book whose rating will be created."
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RatingAdd"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Rating"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Ratings"
                ],
                "summary": "Update rating",
                "operationId": "update_rating",
                "parameters": [
                    {
                        "name": "book_isbn",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Book ISBN",
                            "examples": [
                                "9780439554930"
                            ]
                        },
                        "description": "ISBN of the rating to update."
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RatingUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Rating"
                                    },
                                    "title": "Response Rating Put"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "Ratings"
                ],
                "summary": "Get rating by book ISBN",
                "operationId": "get_rating",
                "parameters": [
                    {
                        "name": "book_isbn",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Book ISBN",
                            "examples": [
                                "9780439554930"
                            ]
                        },
                        "description": "ISBN of the rating to retrieve."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Rating"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Rating not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Ratings"
                ],
                "summary": "Delete Rating",
                "operationId": "delete_rating",
                "parameters": [
                    {
                        "name": "book_isbn",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Book ISBN",
                            "examples": [
                                "9780439554930"
                            ]
                        },
                        "description": "ISBN of the rating to delete."
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "404": {
                        "description": "Rating not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "Rating": {
                "properties": {
                    "book_isbn": {
                        "type": "string",
                        "title": "Book ISBN",
                        "examples": [
                            "9780439554930"
                        ]
                    },
                    "num_ratings": {
                        "type": "integer",
                        "format": "int64",
                        "title": "Number of Ratings",
                        "examples": [
                            9234567
                        ]
                    },
                    "star_rating": {
                        "type": "number",
                        "format": "double",
                        "title": "Star Rating",
                        "minimum": 0,
                        "maximum": 5,
                        "examples": [
                            4.47
                        ]
                    }
                },
                "type": "object",
                "required": [
                    "book_isbn",
                    "num_ratings",
                    "star_rating"
                ],
                "title": "Rating"
            },
...
            "RatingAdd": {
                "properties": {
                    "num_ratings": {
                        "type": "integer",
                        "format": "int64",
                        "title": "Number of Ratings",
                        "examples": [
                            9234567
                        ]
                    },
                    "star_rating": {
                        "type": "number",
                        "format": "double",
                        "title": "Star Rating",
                        "minimum": 0,
                        "maximum": 5,
                        "examples": [
                            4.47
                        ]
                    }
                },
                "type": "object",
                "required": [
                    "num_ratings",
                    "star_rating"
                ],
                "title": "RatingAdd"
            },
            "RatingUpdate": {
                "properties": {
                    "num_ratings": {
                        "type": "integer",
                        "format": "int64",
                        "title": "Number of Ratings",
                        "examples": [
                            9234567
                        ]
                    },
                    "star_rating": {
                        "type": "number",
                        "format": "double",
                        "title": "Star Rating",
                        "minimum": 0,
                        "maximum": 5,
                        "examples": [
                            4.47
                        ]
                    }
                },
                "type": "object",
                "required": [
                    "num_ratings",
                    "star_rating"
                ],
                "title": "RatingUpdate"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        },
...
    }
}